src/RPiPico2TestPlatform/tests/host/requirements.txt
pytest>=7.4
pyserial>=3.5
